requirements.txt
psycopg2.binary==2.9.2
python-dotenv==0.19.2